move determination of log level into support crate
