+++ axum-on-rails/src/util.rs
@@ -2,8 +2,9 @@ use figment::{
     providers::{Format, Toml},
     Figment,
 };
-use serde::de::Deserialize;
 use std::env;
+use serde::de::Deserialize;
+use tracing::Level;
 
 pub enum Environment {
     Development,
@@ -27,10 +28,7 @@ pub fn get_env() -> Environment {
     }
 }
 
-pub fn load_config<'a, T>() -> T
-where
-    T: Deserialize<'a>,
-{
+pub fn load_config<'a, T>() -> T where T: Deserialize<'a> {
     let environment = get_env();
     let env_config_file = match environment {
         Environment::Development => "development.toml",
@@ -48,3 +46,21 @@ where
         .expect("Could not read configuration!");
     config
 }
+
+pub fn get_log_level() -> Level {
+    match env::var("RUST_LOG") {
+        Ok(val) => match val.to_lowercase().as_str() {
+            "trace" => Level::TRACE,
+            "debug" => Level::DEBUG,
+            "info" => Level::INFO,
+            "warn" => Level::WARN,
+            "error" => Level::ERROR,
+            unknown => {
+                eprintln!(r#"Unknown log level: "{}"!"#, unknown);
+                std::process::exit(1)
+            }
+        },
+        Err(_) => Level::INFO,
+    }
+}
+
